Move fast files from feature_85 branch
