+++ wave_vision/include/wave/vision/fast_detector.hpp
@@ -1,6 +1,6 @@
 /**
  * @file
- * Fast Feature Detector implementation, derived from Feature Detector base
+ * FAST Feature Detector implementation, derived from Feature Detector base
  * class.
  * @ingroup vision
  */
@@ -12,7 +12,7 @@
 #include <vector>
 
 /** Libwave Headers */
-#include "wave/vision/feature_detector.hpp"
+#include "wave/vision/detector/feature_detector.hpp"
 
 /** The wave namespace */
 namespace wave {
@@ -44,6 +44,12 @@ struct FASTParams {
      *  0 = TYPE_5_8
      *  1 = TYPE_7_12
      *  2 = TYPE_9_16 (default)
+     *
+     *  The neighbourhood refers to the pixel circumference and number of
+     *  pixels that must be brighter or darker than the center pixel for the
+     *  algorithm to deem the point as a corner. For example, TYPE_9_16
+     *  requires 9 consecutive pixels out of a 16 pixel circumference circle to
+     *  be brighter or darker than the center pixel.
      */
     int type;
 };
