+++ wave_vision/src/fast_detector.cpp
@@ -1,5 +1,5 @@
 // Libwave Headers
-#include "wave/vision/fast_detector.hpp"
+#include "wave/vision/detector/fast_detector.hpp"
 
 namespace wave {
 
@@ -40,7 +40,7 @@ FASTDetector::FASTDetector(const std::string &config_path) {
     parser.addParam("fast.type", &config.type);
 
     if (parser.load(config_path) != 0) {
-        throw ConfigurationLoadingException{};
+        throw std::invalid_argument("Failed to Load Detector Configuration");
     }
 
     // Verify configuration values
